requirements.txt
Flask
gunicorn
sklearn
pandas
numpy
PyYAML
matplotlib
seaborn
evidently
streamlit==1.12.0
kaggle
dill
boto3
- e .


